backend/requirements.txt
Django==5.0.2
# psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-environ==0.11.2
djangorestframework==3.14.0
django-filter==23.5
drf-yasg==1.21.7
djangorestframework-simplejwt==5.3.1
setuptools==69.0.3